feat(entity): cria entity HiringProcess
