+++ ormconfig.json
@@ -1,24 +1,24 @@
 {
-   "type": "postgres",
-   "host": "localhost",
-   "port": 5432,
-   "username": "postgres",
-   "password": "123456",
-   "database": "aceleradora_agil",
-   "synchronize": true,
-   "logging": false,
-   "entities": [
-      "src/models/entity/**/*.ts"
-   ],
-   "migrations": [
-      "src/models/migration/**/*.ts"
-   ],
-   "subscribers": [
-      "src/models/subscriber/**/*.ts"
-   ],
-   "cli": {
-      "entitiesDir": "src/models/entity",
-      "migrationsDir": "src/models/migration",
-      "subscribersDir": "src/models/subscriber"
-   }
+  "type": "postgres",
+  "host": "localhost",
+  "port": 5432,
+  "username": "postgres",
+  "password": "1234",
+  "database": "aceleradora_agil",
+  "synchronize": true,
+  "logging": false,
+  "entities": [
+    "src/models/entity/**/*.ts"
+  ],
+  "migrations": [
+    "src/models/migration/**/*.ts"
+  ],
+  "subscribers": [
+    "src/models/subscriber/**/*.ts"
+  ],
+  "cli": {
+    "entitiesDir": "src/models/entity",
+    "migrationsDir": "src/models/migration",
+    "subscribersDir": "src/models/subscriber"
+  }
 }
